requirements.txt
asgiref==3.8.1
Django==5.1.1
mfrc522==0.0.7
RPi.GPIO==0.7.1
spidev==3.6
sqlparse==0.5.1
